requirements.txt
django
djangorestframework
Pillow
django-rest-knox
django-filter
jinja2
cryptography
boto3